Add mount_omniauth API helper to mount OmniAuthAPI
This is slightly more complicated than the other APIs because devise
token auth has two routes hitting the same controller, so we need to do
a little dance with two Grape::APIs because Grape doesn't really support
aliasing.

Also, have to do some weird stuff in the spec to reload the API after
each test so that you get clean paths. If not you get recursive paths
like '/auth/auth/auth/:provider/callback'. Theres probably a better way
to do this, and it raise the problem that mount_x can't be called twice.