Modify testing setup around OmniAuth
There were a couple of OmniAuth testing problems, notably the
re-requiring of the API file to get a clean test state broke coverage.
It was really necessary after some work to get the tests not to require
it. A few other tiny clean ups to make the spec helper cleaner.