Add secret and key generator in configuration
Has a setting for setting a secret like devise. Additionally, this
commit adds a key generator class, which is copied out of devise. This
allows us to generate a key used for HMAC. This class is largely
untested for now because it comes from devise and just a wrapper around
OpenSSL.