Add flag on Auth Data for how auth was achieved
AuthenticationData now tracks whether authentication happened by token
in a flag variable called 'authed_with_token' which defaults to false.
Therefore, after authing with a token this flag should be toggled to
true. The reason for this change is so that we can switch on how header
creation happens later on in the request.