Fix case_insensitive_key clobbering
Fixes issue #37 returning nil because the AR token auth module can
clobber the Devise method. So, the behavior is now that the reader
method looks to see if an instance variable is set on the class, then it
will use Devise's class variable set on the Devise module. If Devise is
not present it falls back to the [:email] default.